# experimental: optimize function listener context stack