This PR changes the context stack used by function listeners to use a vector instead of a linked list in order to reduce the memory footprint; creating the context stack items forced the compiler to put values on the heap, and the extra pointer increased by 50% the memory space needed to track the contexts (3 pointers per item vs 2 now).

```
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/engine/compiler
                                           │  /tmp/old   │              /tmp/new               │
                                           │   sec/op    │   sec/op     vs base                │
CallEngine_builtinFunctionFunctionListener   37.45n ± 5%   11.10n ± 1%  -70.37% (p=0.000 n=10)

                                           │  /tmp/old  │              /tmp/new              │
                                           │    B/op    │   B/op     vs base                 │
CallEngine_builtinFunctionFunctionListener   24.00 ± 0%   0.00 ± 0%  -100.00% (p=0.000 n=10)

                                           │  /tmp/old  │              /tmp/new               │
                                           │ allocs/op  │ allocs/op   vs base                 │
CallEngine_builtinFunctionFunctionListener   1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
```