@ncruces 2023-05-09
The power of [`cdr` coding](https://en.wikipedia.org/wiki/CDR_coding)… I'll show myself.